+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- **CUMULUS-1579**
+  - Elasticsearch list queries use `match` instead of `term`. `term` had been analyzing the terms and not supporting `-` in the field values.
+
 - **CUMULUS-1619**
   - Adds 4 new keys to `@cumulus/logger` to display granules, parentArn, asyncOperationId, and stackName.
   - Depends on `cumulus-message-adapter-js` version 1.0.10+. Cumulus tasks updated to use this version.
@@ -224,6 +227,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - **CUMULUS-1620** - Fixed bug where `message_adapter_version` does not correctly inject the CMA
+
+- **CUMULUS-1572** - A granule is now included in discovery results even when
+none of its files has a matching file type in the associated collection
+configuration. Previously, if all files for a granule were unmatched by a file
+type configuration, the granule was excluded from the discovery results.
+Further, added support for a `boolean` property
+`ignoreFilesConfigForDiscovery`, which controls how a granule's files are
+filtered at discovery time.
 
 ## [v1.14.2] - 2019-10-08